+++ src/app/profile/page.tsx
@@ -5,7 +5,8 @@ import { BackButton } from "@/components/shared/BackButton";
 import { VerticalCard } from "@/components/shared/VerticalCard";
 import Link from "next/link";
 import { AppContext } from "../../context/AppContextProvider";
-import { CursorResponse, PropertyType, ReviewType, UserSettingsType } from "@/types";
+import { CursorResponse, ReviewType, UserSettingsType } from "@/types";
+import { PropertyType } from "@/types/property";
 import { deleteUserProperty, getUserListedProp } from "@/services/propertyApi";
 import { FaEdit, FaEye, FaTrash } from "react-icons/fa";
 import { toast } from "react-toastify";
