+++ src/components/PropertyDetailsClient.tsx
@@ -6,7 +6,8 @@ import { VerticalCard } from "@/components/shared/VerticalCard";
 import Image from "next/image";
 import { useRouter } from "next/navigation";
 import { FaArrowLeft, FaRegHeart, FaStar } from "react-icons/fa";
-import { PropertyType, ReviewType } from "@/types";
+import { PropertyType } from "@/types/property";
+import { ReviewType } from "@/types";
 import Link from "next/link";
 import { getCollocatedProperties } from "@/services/propertyApi";
 import Price from "@/components/shared/Price";
