+++ src/types.ts
@@ -1,3 +1,5 @@
+import { PropertyType } from "./types/property";
+
 export interface AuthUserType {
   _id: string;
   display_name: string;
@@ -8,33 +10,6 @@ export interface AuthUserType {
   createdAt?: Date;
   updatedAt?: Date
 };
-
-export interface AgentProps { 
-  id: string, 
-  name: string, 
-  rating: number, 
-  sold: number, 
-  image: string,
-  reviews: number, 
-  email: string,
-  listing: number,
-  properties?: PropertyProps[]
-}
-
-export interface PropertyProps {
-  id: string,
-  image: string; // URL of the property image
-  name: string; // Title of the property
-  rating: number; // Rating of the property
-  address: string; // Location of the property
-  price: number; // Price per month
-  currency: CurrencyEnum;
-  category: CategoryEnum; // Property type (e.g., "Apartment")
-  period: string;
-  listed_for: string;
-  distance?: string;
-  expired?: boolean
-}
 
 export type ReviewType = {
   _id: string;
@@ -97,64 +72,6 @@ export interface UserSettingsType {
   social_handles?: {platform: string; handle: string}[]; // e.g [{platform: 'facebook', handle: 'fb.com/username'}]
 };
 
-export enum CurrencyEnum {
-  naira = "NGN",
-  dollar="USD",
-  pound = "GBP",
-  euro = "EUR",
-}
-
-export enum CategoryEnum {
-  house = "house",
-  shortlet="shortlet",
-  hotel = "hotel",
-  office = "office",
-  land = "land",
-  shop = "shop",
-  others = "others"
-}
-
-export enum ListForEnum {
-  rent = "rent",
-  sale = "sale"
-}
-
-export enum RenewalEnum {
-  monthly = "monthly",
-  yearly = "yearly",
-  daily = "daily",
-  weekly = "weekly"
-}
-
-export enum NegotiableEnum {
-  Negotiable = "negotiable",
-  NonNegotiable = "Non-negotiable",
-}
-
-export enum PropertyStatusEnum {
-  available = "available",
-  sold = "sold",
-  rented = "rented",
-  unavailable = "unavailable",
-  expired = "expired",
-}
-
-export interface PropertyFilterPayload {
-  location?: {
-    query: string;
-    lat: number;
-    lng: number;
-    name: string;
-    lga?: string;
-    state?: string;
-  };
-  propertyType: CategoryEnum;
-  listedFor: "all" | "sale" | "rent";
-  priceMin: number | null;
-  priceMax: number | null;
-  description?: string;
-}
-
 export interface Landmark {
   id:       string;
   name:     string;
@@ -167,40 +84,6 @@ export type LandmarkCategory =
   | "school" | "hospital" | "market"  | "transport"
   | "mall"   | "bank"     | "restaurant" | "park"
   | "place_of_worship"    | "other";
-
-export interface PropertyType {
-  _id: string;
-  banner: string; // URL of the property image or image with index = 0
-  title: string; // Title of the property (e.g. 3 bedroom flat, self contain, )
-  slug: string;
-  address: string; // Location of the property
-  price: number; // Price per month
-  currency: CurrencyEnum; // Currency type (e.g. NGN, USD, GBP)
-  listed_for: string; // (e.g. "sell"/ "rent")
-  category: CategoryEnum; // The class of property (e.g. house, land, shop, office, hotel )
-  period?: RenewalEnum; // if is for rent, payment period (e.g monthly, yearly, daily)
-  negotiable: boolean; // (true/false)
-  description?: string // agent's term cond
-  duration?: number;
-  expired_by?: Date;
-  images: string[]; //Other property images for gallery
-  user: UserSettingsType;
-  latitude: number;
-  longitude: number;
-  features?: string[];
-  distance?: string;
-  average_rating?: number;
-  review_count?: number;
-  landmarks?: any[];
-  status: PropertyStatusEnum; // (available, sold, unavailable, rented)
-  createdAt?: Date;
-  updatedAt?: Date;
-};
-
-export interface Feature {
-  name: string;
-  quantity: number;
-}
 
 export interface AgentType { 
   id: string;
